refactor: split off monodeploy-git (#233)
