+++ packages/git/package.json
@@ -7,6 +7,7 @@
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
   },
+  "main": "lib/index.js",
   "files": [
     "lib"
   ],
@@ -16,5 +17,11 @@
     "test:watch": "run workspace:test:watch \"$(pwd)\"",
     "build:watch": "run workspace:build:babel:watch \"$(pwd)\"",
     "prepack": "run workspace:build \"$(pwd)\""
+  },
+  "devDependencies": {
+    "@types/node": "^14.14.25"
+  },
+  "dependencies": {
+    "monodeploy-logging": "workspace:packages/logging"
   }
 }
