+++ packages/node/src/monodeploy.test.ts
@@ -4,14 +4,14 @@ import path from 'path'
 
 import * as npm from '@yarnpkg/plugin-npm'
 
+import * as git from 'monodeploy-git'
 import { LOG_LEVELS } from 'monodeploy-logging'
 
 import monodeploy from './monodeploy'
 import type { MonodeployConfiguration } from './types'
-import * as git from './utils/git'
 
 jest.mock('@yarnpkg/plugin-npm')
-jest.mock('./utils/git')
+jest.mock('monodeploy-git')
 
 const mockGit = git as jest.Mocked<
     typeof git & {
